Isolate log level types
