+++ src/functions/logs/get-log-levels.spec.ts
@@ -6,7 +6,7 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { isValidLogLevel, LogLevel } from '../../models';
+import { isLogLevel, LogLevel } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { makeGetLogLevels } from './get-log-levels';
@@ -18,7 +18,7 @@ describe('getLogLevels()', () => {
 		'Should get all available log levels and the current active log level',
 		integrationTest(async () => {
 			const { available, current } = await getLogLevels(TEST_AUTH_TOKEN);
-			const _isValidLogLevel = (value: any): value is LogLevel | 'off' => value === 'off' || isValidLogLevel(value);
+			const _isValidLogLevel = (value: any): value is LogLevel | 'off' => value === 'off' || isLogLevel(value);
 			expect(_isValidLogLevel(current)).toBeTrue();
 			expect(available.every(_isValidLogLevel)).toBeTrue();
 		}),
