+++ src/models/log-level/log-level-to-raw.ts
@@ -1,0 +1,14 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export const LOG_LEVEL_TO_RAW = {
+	'error': 'Error',
+	'warning': 'Warn',
+	'information': 'Info',
+	'web access': 'Web Access',
+} as const;
